ray AIR for sklearn models
